Fixed merge conflicts and fixed bugs
